requirements.txt
aiofiles>=24.0.0
aiohttp>=3.9.0
asyncpg>=0.30.0
evidently>=0.4.0
fakeredis>=2.25.0
fastapi>=0.110.0
greenlet>=3.2.3
hdbscan>=0.8.29
jsonschema>=4.20.0
lz4>=4.0.0
mcp-context-sdk>=0.1.0
mcp>=1.10.1
mlflow>=2.9.0
mlxtend>=0.23.0
nltk>=3.8.0
numpy>=1.24.0
optuna>=3.5.0
orjson>=3.9.0
pandas>=2.0.0
prometheus-client>=0.19.0
psycopg_pool>=3.1.0
psycopg[binary]>=3.1.0
pydantic-settings>=2.10.1
pydantic>=2.5.0
pytest-benchmark>=4.0.0
pyyaml>=6.0.0
redis>=5.0.0
rich>=13.7.0
scikit-learn>=1.4.0
scipy>=1.10.0
sentence-transformers>=2.5.0
sqlmodel>=0.0.24
textstat>=0.7.0
textual>=0.80.0
transformers>=4.30.0
uvicorn[standard]>=0.24.0
uvloop>=0.19.0
websockets>=12.0
